requirements.txt
hid==1.0.5
Pillow==9.3.0
requests==2.28.2